+++ textpy/utils/auto_docstring_custom.mustache
@@ -1,8 +1,6 @@
 {{#placeholder}}
 {{/placeholder}}
 {{summaryPlaceholder}}.
-
-{{extendedSummaryPlaceholder}}
 {{#parametersExist}}
 
 Parameters
